requirements.txt
# FastAPI & Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.7

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# AWS SDK & Storage
boto3>=1.29.0,<1.35.0
aioboto3>=11.3.0
Pillow==10.1.0

# HTTP Client for Midtrans
httpx==0.25.2
aiofiles==23.2.1

# Validation & Schemas
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0

# Linting & Formatting
ruff==0.1.6
black==23.11.0
isort==5.12.0

# Utilities
loguru==0.7.2
python-slugify==8.0.1